requirements.txt
flask
flask-mysqldb
numpy
